Djangoの設定ファイルを開発用と本番用で追記修正
